+++ qom/ui/plotters/MPLPlotter.py
@@ -6,7 +6,7 @@
 __name__    = 'qom.ui.plotters.MPLPlotter'
 __authors__ = ['Sampreet Kalita']
 __created__ = '2020-10-03'
-__updated__ = '2021-02-10'
+__updated__ = '2021-02-24'
 
 # dependencies
 from matplotlib.colors import Normalize
@@ -29,8 +29,7 @@ t_list = Union[list, np.ndarray]
 
 # TODO: Add annotations.
 # TODO: Options for 3D plot parameters.
-# TODO: Options for lower and upper bounds.
-# TODO: Options for `ticklabel_format`.
+# TODO: Options for `ticklabel_format` and padding.
 
 class MPLPlotter(BasePlotter):
     """Class to handle matplotlib plots.
@@ -56,73 +55,45 @@ class MPLPlotter(BasePlotter):
         _mpl_axes = plt.gca(projection='3d' if _type in self.types_3D else None)
         _font_dicts = self.params['font_dicts']
 
-        # update math fonts
+        # update fonts
         plt.rcParams['mathtext.fontset'] = _font_dicts['math']
 
         # update title
-        plt.title(self.params['title'], fontdict=_font_dicts['label'])
-        
-        # update labels
-        plt.xlabel(self.axes['X'].label, labelpad=12, fontdict=_font_dicts['label'])
+        _mpl_axes.set_title(self.params['title'], fontdict=_font_dicts['label'], pad=12)
 
-        # update tick properties
-        _font_props = self.__get_font_props(_font_dicts['tick'])
-        plt.setp(_mpl_axes.get_xticklabels(), fontproperties=_font_props)
-        plt.setp(_mpl_axes.get_yticklabels(), fontproperties=_font_props)
+        # udpate ticks
         plt.ticklabel_format(axis='both', style='plain')
 
-        # update ticks
-        _mpl_axes.set_xlim(min(self.axes['X'].ticks), max(self.axes['X'].ticks))
-        _mpl_axes.set_xticks(self.axes['X'].ticks)
-        _mpl_axes.set_xticklabels(self.axes['X'].tick_labels)
+        # update x-axis
+        self.__update_axis(_mpl_axes, 'x', self.axes['X'])
 
         # 1D plot
         if _type in self.types_1D:
-            # update labels
-            plt.ylabel(self.axes['V'].label, labelpad=12, fontdict=_font_dicts['label'])
-
-            # update ticks
-            _mpl_axes.set_ylim(min(self.axes['V'].ticks), max(self.axes['V'].ticks))
-            _mpl_axes.set_yticks(self.axes['V'].ticks)
-            _mpl_axes.set_yticklabels(self.axes['V'].tick_labels)
+            # update y-axis
+            self.__update_axis(_mpl_axes, 'y', self.axes['V'])
 
             # initialize 1D plot
             self.__init_1D()
 
         # 2D plot
         elif _type in self.types_2D:
-            # update labels
-            plt.ylabel(self.axes['Y'].label, labelpad=12, fontdict=_font_dicts['label'])
-
-            # update ticks
-            _mpl_axes.set_ylim(min(self.axes['Y'].ticks), max(self.axes['Y'].ticks))
-            _mpl_axes.set_yticks(self.axes['Y'].ticks)
-            _mpl_axes.set_yticklabels(self.axes['Y'].tick_labels)
+            # update y-axis
+            self.__update_axis(_mpl_axes, 'y', self.axes['Y'])
 
             # initializze 2D plot
             self.__init_2D()
 
         # 3D plot
         else:
-            # update labels
-            plt.ylabel(self.axes['Y'].label, labelpad=12, fontdict=_font_dicts['label'])
-            _mpl_axes.set_zlabel(self.axes['Z'].label, labelpad=12, fontdict=_font_dicts['label'])
-            
-            # update tick properties
-            plt.setp(_mpl_axes.get_zticklabels(), fontproperties=_font_props)
-
-            # update ticks
-            _mpl_axes.set_ylim(min(self.axes['Y'].ticks), max(self.axes['Y'].ticks))
-            _mpl_axes.set_yticks(self.axes['Y'].ticks)
-            _mpl_axes.set_yticklabels(self.axes['Y'].tick_labels)
-            _mpl_axes.set_zlim(min(self.axes['V'].ticks), max(self.axes['V'].ticks))
-            _mpl_axes.set_zticks(self.axes['V'].ticks)
-            _mpl_axes.set_zticklabels(self.axes['V'].tick_labels)
+            # update y-axis
+            self.__update_axis(_mpl_axes, 'y', self.axes['Y'])
+            # update z-axis
+            self.__update_axis(_mpl_axes, 'z', self.axes['V'])
 
             # initializze 3D plot
             self.__init_3D()
 
-    def __get_font_props(self, font_dict):
+    def __get_font_props(self, font_dict: dict):
         """Method to convert font dictionary to FontProperties.
          
         Parameters
@@ -154,32 +125,33 @@ class MPLPlotter(BasePlotter):
 
         # extract frequently used variables
         _type = self.params['type']
-        palette = self.params['palette']
         _mpl_axes = plt.gca(projection=None)
+        _palette = self.params['palette']
+        _colors = self.axes['Y'].colors
+        _styles = self.axes['Y'].styles
         _dim = len(self.axes['Y'].legends)
 
         # handle dimension for single-value plot
         if _type == 'line' or _type == 'scatter':
             _dim = 1
 
-        # colors
-        _colors = self.axes['Y'].colors
+        # udpate colors
         if _colors is None:
-            _colors = self.color_palettes(palette, self.bins, False) if palette in self.default_palettes else self.diverging_palettes.get(palette, 'blr')(self.bins, False)
+            _colors = self.get_palette(_palette, self.bins, False)
             self.axes['Y'].colors = _colors
 
         # line plots
-        if _type == 'line' or _type == 'lines':
-            # collection
-            self.plot = [Line2D([], [], color=_colors[i], linestyle=self.axes['Y'].styles[i]) for i in range(_dim)]
-            [_mpl_axes.add_line(self.plot[i]) for i in range(_dim)]
+        if 'line' in _type:
+            # plots
+            self.plots = [Line2D([], [], color=_colors[i], linestyle=_styles[i]) for i in range(_dim)]
+            [_mpl_axes.add_line(self.plots[i]) for i in range(_dim)]
             # heads
-            self.head = [Line2D([], [], color=_colors[i], linestyle=self.axes['Y'].styles[i], marker='o') for i in range(_dim)]
-            [_mpl_axes.add_line(self.head[i]) for i in range(_dim)]
+            self.heads = [Line2D([], [], color=_colors[i], linestyle=_styles[i], marker='o') for i in range(_dim)]
+            [_mpl_axes.add_line(self.heads[i]) for i in range(_dim)]
 
         # scatter plots
-        elif _type == 'scatter' or _type == 'scatters':
-            self.plot = [_mpl_axes.scatter([], [], c=_colors[i], s=self.axes['Y'].sizes[i], marker=self.axes['Y'].styles[i]) for i in range(_dim)]
+        elif 'scatter' in _type:
+            self.plots = [_mpl_axes.scatter([], [], c=_colors[i], s=self.axes['Y'].sizes[i], marker=_styles[i]) for i in range(_dim)]
 
         # legends
         if self.params['legend']['show']:
@@ -192,8 +164,8 @@ class MPLPlotter(BasePlotter):
         # extract frequently used variables
         _type = self.params['type']
         _mpl_axes = plt.gca(projection=None)
-        _cmap = self.params['cmap']
         _font_dicts = self.params['font_dicts']
+        _cmap = self.params['cmap']
 
         # initailize values
         _xs, _ys = np.meshgrid(self.axes['X'].val, self.axes['Y'].val)
@@ -203,45 +175,18 @@ class MPLPlotter(BasePlotter):
 
         # contour plot
         if _type == 'contour':
-            self.plot = _mpl_axes.contour(_xs, _ys, _zeros, levels=self.bins, cmap=self.params['cmap'])
+            self.plots = _mpl_axes.contour(_xs, _ys, _zeros, levels=self.bins, cmap=_cmap)
 
         # contourf plot
         if _type == 'contourf':
-            self.plot = _mpl_axes.contourf(_xs, _ys, _zeros, levels=self.bins, cmap=_cmap)
+            self.plots = _mpl_axes.contourf(_xs, _ys, _zeros, levels=self.bins, cmap=_cmap)
 
         # pcolormesh plot
         if _type == 'pcolormesh':
-            self.plot = _mpl_axes.pcolormesh(_xs, _ys, _nan, shading='gouraud', cmap=_cmap)
+            self.plots = _mpl_axes.pcolormesh(_xs, _ys, _nan, shading='gouraud', cmap=_cmap)
 
         # color bar
-        if self.params['cbar']['show']:
-            # handle type of plot
-            if 'contour' in _type:
-                _sm = plt.cm.ScalarMappable(cmap=_cmap, norm=Normalize(vmin=0, vmax=0))
-                _sm.set_array([])
-            else:
-                _sm = self.plot
-            self.cbar = plt.colorbar(_sm)
-
-            # set title
-            self.cbar.ax.set_title(self.params['cbar']['title'], fontproperties=self.__get_font_props(_font_dicts['label']))
-            # set labels
-            self.cbar.ax.set_xlabel(self.params['cbar']['x_label'], labelpad=_font_dicts['tick']['size'] + 12, fontproperties=self.__get_font_props(_font_dicts['label']))
-            self.cbar.ax.set_ylabel(self.params['cbar']['y_label'], labelpad=_font_dicts['tick']['size'] + 12, fontproperties=self.__get_font_props(_font_dicts['label']))
-            # set ticks
-            _ticks = self.params['cbar']['ticks']
-            if _ticks is None:
-                _mini, _maxi, _prec = super().get_limits(0.0, 1.0)
-                _ticks = np.around(np.linspace(_mini, _maxi, self.bins), decimals=_prec)
-            self.cbar.set_ticks(_ticks)
-            # set tick labels
-            _tick_labels = self.params['cbar']['tick_labels']
-            if _tick_labels is None:
-                _tick_labels = _ticks
-            self.cbar.set_ticklabels(_tick_labels)
-            plt.setp(self.cbar.ax.get_yticklabels(), fontproperties=self.__get_font_props(_font_dicts['tick']))
-        else:
-            self.cbar = None
+        self.cbar = None
 
     def __init_3D(self):
         """Method to initialize 3D plots."""
@@ -250,7 +195,6 @@ class MPLPlotter(BasePlotter):
         _type = self.params['type']
         _mpl_axes = plt.gca(projection='3d')
         _cmap = self.params['cmap']
-        _font_dicts = self.params['font_dicts']
 
         # update view
         _mpl_axes.view_init(32, 216)
@@ -274,39 +218,19 @@ class MPLPlotter(BasePlotter):
 
         # surface plot
         if 'surface' in _type:
-            self.plot =_mpl_axes.plot_surface(_xs, _ys, _zeros, rstride=1, cstride=1, cmap=_cmap)
+            self.plots =_mpl_axes.plot_surface(_xs, _ys, _zeros, rstride=1, cstride=1, cmap=_cmap)
 
-        # add color bar
-        if self.params['cbar']['show']:
-            self.cbar = plt.colorbar(self.plot)
-            # set title
-            self.cbar.ax.set_title(self.params['cbar']['title'], fontproperties=self.__get_font_props(_font_dicts['label']))
-            # set labels
-            self.cbar.ax.set_xlabel(self.params['cbar']['x_label'], labelpad=_font_dicts['tick']['size'] + 12, fontproperties=self.__get_font_props(_font_dicts['label']))
-            self.cbar.ax.set_ylabel(self.params['cbar']['y_label'], labelpad=_font_dicts['tick']['size'] + 12, fontproperties=self.__get_font_props(_font_dicts['label']))
-            # set ticks
-            _ticks = self.params['cbar']['ticks']
-            if _ticks is None:
-                _mini, _maxi, _prec = super().get_limits(0.0, 1.0)
-                _ticks = np.around(np.linspace(_mini, _maxi, self.bins), decimals=_prec)
-            self.cbar.set_ticks(_ticks)
-            # set tick labels
-            _tick_labels = self.params['cbar']['tick_labels']
-            if _tick_labels is None:
-                _tick_labels = _ticks
-            self.cbar.set_ticklabels(_tick_labels)
-            plt.setp(self.cbar.ax.get_yticklabels(), fontproperties=self.__get_font_props(_font_dicts['tick']))
-        else:
-            self.cbar = None
+        # color bar
+        self.cbar = None
 
-    def __update_1D(self, xs, ys, head):
+    def __update_1D(self, xs: t_list, vs: t_list, head: bool):
         """Method to udpate 1D plots.
         
         Parameters
         ----------
         xs : list or numpy.ndarray
             X-axis values.
-        ys : list or numpy.ndarray
+        vs : list or numpy.ndarray
             Y-axis values.
         head : boolean
             Option to display the head for line-type plots.
@@ -317,43 +241,44 @@ class MPLPlotter(BasePlotter):
         _mpl_axes = plt.gca(projection=None)
         _dim = len(xs[0])
         
-        # update line plots
-        if _type == 'line' or _type == 'lines':
-            for j in range(len(self.plot)):
-                self.plot[j].set_xdata(xs[j])
-                self.plot[j].set_ydata(ys[j])
-                _idx_nan = np.argwhere(np.isnan(ys[j]))
+        # line plots
+        if 'line' in _type:
+            for j in range(len(self.plots)):
+                self.plots[j].set_xdata(xs[j])
+                self.plots[j].set_ydata(vs[j])
+                _idx_nan = np.argwhere(np.isnan(vs[j]))
                 _idx_nan = _idx_nan[0][0] if len(_idx_nan) != 0 else -1
                 if head and _idx_nan < _dim - 1 and _idx_nan != -1:
-                    self.head[j].set_xdata(xs[j][_idx_nan - 1 : _idx_nan])
-                    self.head[j].set_ydata(ys[j][_idx_nan - 1 : _idx_nan])
+                    self.heads[j].set_xdata(xs[j][_idx_nan - 1 : _idx_nan])
+                    self.heads[j].set_ydata(vs[j][_idx_nan - 1 : _idx_nan])
                 else:
-                    self.head[j].set_xdata([])
-                    self.head[j].set_ydata([])
+                    self.heads[j].set_xdata([])
+                    self.heads[j].set_ydata([])
 
-        # update scatter plots
-        if _type == 'scatter' or _type == 'scatters':
-            for j in range(len(self.plot)):
-                XY = np.c_[xs[j], ys[j]]
-                self.plot[j].set_offsets(XY)
+        # scatter plots
+        if 'scatter' in _type:
+            for j in range(len(self.plots)):
+                XY = np.c_[xs[j], vs[j]]
+                self.plots[j].set_offsets(XY)
                 
         # handle nan values for limits
         _minis = []
         _maxis = []
-        for j in range(len(ys)):
+        for j in range(len(vs)):
             # calculate minimum and maximum values
-            if len(ys[j]) != 0:
+            if len(vs[j]) != 0:
                 # handle NaN values
-                _no_nan = [y if y == y else 0 for y in ys[j]]
+                _no_nan = [y if y == y else 0 for y in vs[j]]
 
                 # update limits
                 _minis.append(min(_no_nan))
                 _maxis.append(max(_no_nan))
 
-        # set limits
+        # get limits
         _mini, _maxi = min(_minis), max(_maxis)
         _mini, _maxi, _prec = super().get_limits(_mini, _maxi)
-        # update ticks and tick labels
+
+        # ticks and tick labels
         _ticks = self.axes['V'].ticks
         _tick_labels = self.axes['V'].tick_labels
         if self.axes['V'].bound == 'none':
@@ -361,11 +286,13 @@ class MPLPlotter(BasePlotter):
             _tick_labels = _ticks
         _mini = min(_ticks)
         _maxi = max(_ticks)
-        _mpl_axes.set_ylim(_mini, _maxi)
         _mpl_axes.set_yticks(_ticks)
         _mpl_axes.set_yticklabels(_tick_labels)
 
-    def __update_2D(self, vs):
+        # update limits
+        _mpl_axes.set_ylim(_mini, _maxi)
+
+    def __update_2D(self, vs: t_list):
         """Method to udpate 2D plots.
         
         Parameters
@@ -377,91 +304,73 @@ class MPLPlotter(BasePlotter):
         # frequently used variables
         _type = self.params['type']
         _mpl_axes = plt.gca(projection=None)
-        _cmap = self.params['cmap']
         _font_dicts = self.params['font_dicts']
+        _cmap = self.params['cmap']
         _rave = np.ravel(vs)
 
         # initialize values
         _no_nan = [z if z == z else 0 for z in _rave]
         _mini, _maxi = min(_no_nan), max(_no_nan)
-        _mini, _maxi, _prec = super().get_limits(_mini, _maxi)
+
+        # handle color bar limits
+        if self.params['cbar']['show'] and self.params['cbar']['ticks'] is not None:
+            _mini = self.params['cbar']['ticks'][0]
+            _maxi = self.params['cbar']['ticks'][-1]
+            _prec = np.ceil(-np.log10(_mini))
+        else:
+            _mini, _maxi, _prec = super().get_limits(_mini, _maxi)
 
         # contour and contourf plots
         if 'contour' in _type:
             # remove QuadContourSet PathCollection
-            for pc in self.plot.collections:
+            for pc in self.plots.collections:
                 pc.remove()
             _xs, _ys = np.meshgrid(self.axes['X'].val, self.axes['Y'].val)
 
             # contour plot
             if _type == 'contour':
-                self.plot = _mpl_axes.contour(_xs, _ys, vs, levels=self.bins, cmap=_cmap)
+                self.plots = _mpl_axes.contour(_xs, _ys, vs, levels=self.bins, cmap=_cmap)
             # contourf plot
             if _type == 'contourf':
-                self.plot = _mpl_axes.contourf(_xs, _ys, vs, levels=self.bins, cmap=_cmap)
-
-            # redraw color bar
-            if self.cbar is not None:
-                self.cbar.ax.clear()
-                _sm = plt.cm.ScalarMappable(cmap=_cmap, norm=Normalize(vmin=_mini, vmax=_maxi))
-                _sm.set_array([])
-                self.cbar = plt.colorbar(_sm, cax=self.cbar.ax)
-
-                # update title
-                self.cbar.ax.set_title(self.params['cbar']['title'], fontproperties=self.__get_font_props(_font_dicts['label']))
-                # update labels
-                self.cbar.ax.set_xlabel(self.params['cbar']['x_label'], labelpad=_font_dicts['tick']['size'] + 12, fontproperties=self.__get_font_props(_font_dicts['label']))
-                self.cbar.ax.set_ylabel(self.params['cbar']['y_label'], labelpad=_font_dicts['tick']['size'] + 12, fontproperties=self.__get_font_props(_font_dicts['label']))
-                # update ticks
-                plt.setp(self.cbar.ax.get_yticklabels(), fontproperties=self.__get_font_props(_font_dicts['tick']))
+                self.plots = _mpl_axes.contourf(_xs, _ys, vs, levels=self.bins, cmap=_cmap)
 
         # pcolormesh plot
         if _type == 'pcolormesh':
-            self.plot.set_array(_rave)
+            self.plots.set_array(_rave)
 
         # set limits
-        self.plot.set_clim(vmin=_mini, vmax=_maxi)
+        self.plots.set_clim(vmin=_mini, vmax=_maxi)
 
         # color bar
-        if self.cbar is not None:
-            # update ticks
-            _ticks = self.params['cbar']['ticks']
-            if _ticks is None:
-                _ticks = np.around(np.linspace(_mini, _maxi, self.bins), decimals=_prec)
-            self.cbar.set_ticks(_ticks)
-            # update tick labels
-            _tick_labels = self.params['cbar']['tick_labels']
-            if _tick_labels is None:
-                _tick_labels = _ticks
-            self.cbar.set_ticklabels(_tick_labels)
-            self.cbar.draw_all()
+        if self.params['cbar']['show']:
+            self.set_cbar(_mini, _maxi, _prec)
 
-    def __update_3D(self, zs):
+    def __update_3D(self, vs: t_list):
         """Method to udpate 3D plots.
         
         Parameters
         ----------
-        zs : list or numpy.ndarray
-            Z-axis values.
+        vs : list or numpy.ndarray
+            Z-axis or V-axis values.
         """
 
         # frequently used variables
         _type = self.params['type']
         _mpl_axes = plt.gca(projection='3d')
-        _cmap = self.params['cmap']
         _font_dicts = self.params['font_dicts']
+        _cmap = self.params['cmap']
 
         # initialize values
         _X, _Y = np.meshgrid(self.axes['X'].val, self.axes['Y'].val)
-        _Z = np.array(zs) if type(zs) is list else zs
+        _Z = np.array(vs) if type(vs) is list else vs
         _mini, _maxi = min(_Z.ravel()), max(_Z.ravel())
         _mini, _maxi, _prec = super().get_limits(_mini, _maxi)
 
         # surface plot
         if 'surface'in _type:
-            _cmap = self.plot.get_cmap()
-            self.plot.remove()
-            self.plot = _mpl_axes.plot_surface(_X, _Y, _Z, rstride=1, cstride=1, cmap=_cmap, antialiased=False)
+            _cmap = self.plots.get_cmap()
+            self.plots.remove()
+            self.plots = _mpl_axes.plot_surface(_X, _Y, _Z, rstride=1, cstride=1, cmap=_cmap, antialiased=False)
 
             # projections
             if _type == 'surface_cz':
@@ -470,7 +379,7 @@ class MPLPlotter(BasePlotter):
 
             # update colorbar
             if self.cbar is not None:
-                self.cbar.update_normal(self.plot)
+                self.cbar.update_normal(self.plots)
 
         # update ticks and tick labels
         _ticks = self.axes['V'].ticks
@@ -482,29 +391,50 @@ class MPLPlotter(BasePlotter):
         _maxi = max(_ticks)
         _mpl_axes.set_zticks(_ticks)
         _mpl_axes.set_zticklabels(_tick_labels)
+
         # update limits
-        self.plot.set_clim(vmin=_mini, vmax=_maxi)
+        self.plots.set_clim(vmin=_mini, vmax=_maxi)
         _mpl_axes.set_zlim(_mini, _maxi)
 
         # color bar
-        if self.cbar is not None:
-            _ticks = self.params['cbar']['ticks']
-            if _ticks is None:
-                _ticks = np.around(np.linspace(_mini, _maxi, self.bins), decimals=_prec)
-            self.cbar.set_ticks(_ticks)
-            # update tick labels
-            _tick_labels = self.params['cbar']['tick_labels']
-            if _tick_labels is None:
-                _tick_labels = _ticks
-            self.cbar.set_ticklabels(_tick_labels)
+        if self.params['cbar']['show']:
+            self.set_cbar(_mini, _maxi, _prec)
             self.cbar.ax.set_autoscale_on(True)
-            self.cbar.draw_all()
+
+    def __update_axis(self, ax, ax_name: str, ax_data: dict):
+        """Method to update an axis.
+        
+        Parameters
+        ----------
+        ax : :class:`matplotlib.axes.Axes`
+            Axis of the plot.
+        ax_name : str
+            Name of the axis.
+        ax_data : dict
+            Data for the axis.
+        """
+
+        # frequently used variables
+        _font_dicts = self.params['font_dicts']
+        _font_props = self.__get_font_props(_font_dicts['tick'])
+
+        # labels
+        getattr(ax, 'set_' + ax_name + 'label')(ax_data.label, labelpad=12, fontdict=_font_dicts['label'])
+
+        # ticks
+        getattr(ax, 'set_' + ax_name + 'lim')(min(ax_data.ticks), max(ax_data.ticks))
+        getattr(ax, 'set_' + ax_name + 'ticks')(ax_data.ticks)
+        ax.tick_params(axis=ax_name, which='major', pad=12)
+
+        # tick labels
+        getattr(ax, 'set_' + ax_name + 'ticklabels')(ax_data.tick_labels)
+        plt.setp(getattr(ax, 'get_' + ax_name + 'ticklabels')(), fontproperties=_font_props)
 
     def gca(self):
         """Method to obtain the axes of the figure.
         
         Returns
-        ----------
+        -------
         axes : :class:`matplotlib.axes.Axes`
             Axes of the figure.
         """
@@ -514,8 +444,104 @@ class MPLPlotter(BasePlotter):
 
         return plt.gca(projection='3d' if _type in self.types_3D else None)
 
+    def gcf(self):
+        """Method to obtain the current figure.
+        
+        Returns
+        -------
+        fig : :class:`matplotlib.pyplot.figure`
+            Current figure.
+        """
+        
+        return plt.gcf()
+
+    def gta(self):
+        """Method to obtain a twin axes of the figure.
+        
+        Returns
+        ----------
+        axes : :class:`matplotlib.axes.Axes`
+            Axes of the figure.
+        """
+
+        # frequently used variables
+        _type = self.params['type']
+        _font_dicts = self.params['font_dicts']
+        _font_props = self.__get_font_props(_font_dicts['tick'])
+
+        # initialize twin axis
+        ax_twin = plt.twinx()
+
+        # labels
+        ax_twin.set_ylabel('a', labelpad=12, fontdict=_font_dicts['label'])
+
+        # ticks
+        plt.setp(ax_twin.get_yticklabels(), fontproperties=_font_props)
+        ax_twin.tick_params(axis='y', which='major', pad=12)
+
+        return ax_twin
+
+    def set_cbar(self, mini: float=0, maxi: float=0, prec: int=2):
+        """Method to set the colorbar for the figure.
+
+        Parameters
+        ----------
+        mini : float
+            Minimum value displayed. Default is 0.
+        maxi : float
+            Maximum value displayed. Default is 0.
+        prec : int
+            Precision upto which the values are displayed. Default is 2.
+        """
+
+        # frequently used variables
+        _orientation = self.params['cbar']['orientation']
+        _font_dicts = self.params['font_dicts']
+        _cax = None
+
+        # clear if existed
+        if self.cbar is not None:
+            self.cbar.ax.clear()
+            _cax = self.cbar.ax
+
+        # set scalar mappable 
+        if 'contour' in self.params['type']:
+            _sm = plt.cm.ScalarMappable(cmap=self.params['cmap'], norm=Normalize(vmin=mini, vmax=maxi))
+            _sm.set_array([])
+        else:
+            _sm = self.plots
+
+        # initialize
+        self.cbar = plt.colorbar(_sm, cax=_cax, orientation=_orientation)
+
+        # title
+        self.cbar.ax.set_title(self.params['cbar']['title'], fontproperties=self.__get_font_props(_font_dicts['label']), pad=12)
+
+        # labels
+        self.cbar.ax.set_xlabel(self.params['cbar']['x_label'], labelpad=_font_dicts['tick']['size'] + 12, fontproperties=self.__get_font_props(_font_dicts['label']))
+        self.cbar.ax.set_ylabel(self.params['cbar']['y_label'], labelpad=_font_dicts['tick']['size'] + 12, fontproperties=self.__get_font_props(_font_dicts['label']))
+
+        # ticks
+        _ticks = self.params['cbar']['ticks']
+        if _ticks is None:
+            _ticks = np.around(np.linspace(mini, maxi, self.bins), decimals=prec)
+        self.cbar.set_ticks(_ticks)
+
+        # tick labels
+        _tick_labels = self.params['cbar']['tick_labels']
+        if _tick_labels is None:
+            _tick_labels = _ticks
+        self.cbar.set_ticklabels(_tick_labels)
+        plt.setp(self.cbar.ax.get_yticklabels(), fontproperties=self.__get_font_props(_font_dicts['tick']))
+        plt.setp(self.cbar.ax.get_xticklabels(), fontproperties=self.__get_font_props(_font_dicts['tick']))
+        self.cbar.ax.tick_params(axis='x', which='major', pad=12)
+        self.cbar.ax.tick_params(axis='y', which='major', pad=12)
+
+        # draw colorbar
+        self.cbar.draw_all()
+
     def update(self, xs: t_list=None, ys: t_list=None, zs: t_list=None, vs: t_list=None, head: bool=False):
-        """Method to update plot.
+        """Method to update the figure.
         
         Parameters
         ----------
@@ -550,12 +576,12 @@ class MPLPlotter(BasePlotter):
             self.__update_3D(vs)
 
     def show(self, hold: bool=False):
-        """Method to display plot.
+        """Method to display the figure.
 
         Parameters
         ----------
         hold : bool, optional
-            Option to hold the plot. Default is True.
+            Option to hold the plot. Default is False.
         """
 
         # draw data